+++ packages/contracts/test/core/permission/permission-manager.ts
@@ -151,12 +151,7 @@ describe('Core: PermissionManager', function () {
           .grant(pm.address, otherSigner.address, ADMIN_PERMISSION_ID)
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          otherSigner.address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, otherSigner.address, ROOT_PERMISSION_ID);
     });
 
     it('should not allow for non ROOT', async () => {
@@ -167,12 +162,7 @@ describe('Core: PermissionManager', function () {
           .grant(pm.address, otherSigner.address, ROOT_PERMISSION_ID)
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          otherSigner.address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, otherSigner.address, ROOT_PERMISSION_ID);
     });
   });
 
@@ -310,12 +300,7 @@ describe('Core: PermissionManager', function () {
           )
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          otherSigner.address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, otherSigner.address, ROOT_PERMISSION_ID);
     });
 
     it('should not allow for non ROOT', async () => {
@@ -336,12 +321,7 @@ describe('Core: PermissionManager', function () {
           )
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          otherSigner.address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, otherSigner.address, ROOT_PERMISSION_ID);
     });
   });
 
@@ -372,12 +352,7 @@ describe('Core: PermissionManager', function () {
           .revoke(pm.address, otherSigner.address, ADMIN_PERMISSION_ID)
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          otherSigner.address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, otherSigner.address, ROOT_PERMISSION_ID);
     });
 
     it('should not emit revoked if already revoked', async () => {
@@ -395,12 +370,7 @@ describe('Core: PermissionManager', function () {
           .revoke(pm.address, otherSigner.address, ADMIN_PERMISSION_ID)
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          otherSigner.address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, otherSigner.address, ROOT_PERMISSION_ID);
     });
 
     it('should not allow for non ROOT', async () => {
@@ -411,12 +381,7 @@ describe('Core: PermissionManager', function () {
           .revoke(pm.address, otherSigner.address, ADMIN_PERMISSION_ID)
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          otherSigner.address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, otherSigner.address, ROOT_PERMISSION_ID);
     });
   });
 
@@ -537,12 +502,7 @@ describe('Core: PermissionManager', function () {
         pm.connect(signers[2]).applyMultiTargetPermissions(bulkItems)
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          signers[2].address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, signers[2].address, ROOT_PERMISSION_ID);
     });
   });
 
@@ -698,12 +658,7 @@ describe('Core: PermissionManager', function () {
           .applySingleTargetPermissions(pm.address, bulkItems)
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          otherSigner.address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, otherSigner.address, ROOT_PERMISSION_ID);
     });
 
     it('should not allow for non ROOT', async () => {
@@ -721,12 +676,7 @@ describe('Core: PermissionManager', function () {
           .applySingleTargetPermissions(pm.address, bulkItems)
       )
         .to.be.revertedWithCustomError(pm, 'Unauthorized')
-        .withArgs(
-          pm.address,
-          pm.address,
-          otherSigner.address,
-          ROOT_PERMISSION_ID
-        );
+        .withArgs(pm.address, otherSigner.address, ROOT_PERMISSION_ID);
     });
   });