+++ packages/contracts/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Refactored the `auth` modifier to always use `where = address(this)` and adapted errors.
 - Removed the `_auth` in `PermissionManager` that allowed for having the `ROOT_PERMISSION_ID` permission for a specific `where` target contract.
 - Removed the `WITHDRAW_PERMISSION_ID`.
 - Removed `DaoAuthorizableCloneable` and `DaoAuthorizableBase`.